tests/requirements.txt
codingame
python-dotenv
